Added reverse operation and tests
